refactor(stella-tui): split the transcript formatters out of model.rs, which had reached its ceiling
model.rs sat at 1500 of the guard's 1500 lines with no baseline entry, so the
next line added to it would have failed the gate. The line count is the
trigger; the reason to split there is that the two halves grow on different
clocks.

SessionModel::apply matches AgentEvent exhaustively and is one of the four
compile-enforced stops when a variant is added, so every new event costs this
file at least a line and the arm cannot be declined. The trailing formatters —
compact_json, cap_middle, format_tool_input, summarize and friends — are
functions over borrowed strings and serde_json::Value with no knowledge of
events at all, and have no reason to grow when the vocabulary does. Keeping
them here meant the vocabulary's growth was paying rent on the formatter's
length, and the ratchet was about to make an ordinary protocol change
impossible.

A pure move: every function is byte-identical apart from the pub(super) a
sibling module needs, and model.rs re-imports them so no call site changed.
Same shape as deck/classify.rs, extracted from deck.rs for the same reason.

No baseline entry was added — the split relieves the ratchet rather than
raising it, which is the rule. 1500 -> 1275, with 225 lines of room for the
fold to keep absorbing variants.

Verified behaviourally by what did NOT change: the 13 deck render snapshots
pass unblessed, so not one rendered cell moved.

Closes #2958
