+++ crates/stella-tui/src/model.rs
@@ -29,10 +29,19 @@ use std::collections::VecDeque;
 mod error_rows;
 pub mod file_state;
 pub mod recall;
+mod summarize;
+
 #[cfg(test)]
 pub use file_state::DIFF_HISTORY;
 pub use file_state::{FileState, MAX_TRACKED_FILES, RememberedDiff};
 pub use recall::{RecallBudget, RecalledFrameRow};
+// Re-imported rather than left qualified, so the split was a pure move: every
+// call site in the fold reads exactly as it did before (#2958). See
+// `summarize`'s module doc for why the seam is there and not elsewhere.
+use summarize::{
+    INPUT_BUDGET, OUTPUT_BUDGET, cap_input_json, cap_middle, format_tool_input, is_file_mutation,
+    summarize, tool_input_path,
+};
 
 /// How many characters of a tool input / output summary we retain on a
 /// transcript line before eliding — the full payload is never needed on the
@@ -1260,240 +1269,6 @@ impl SessionModel {
             self.files.push(state);
         }
     }
-}
-
-/// Compact a tool-call input `Value` to a single-line JSON string. Falls back
-/// to the empty string on the (impossible for `Value`) serialization error so
-/// the model never panics on a tool card.
-fn compact_json(value: &serde_json::Value) -> String {
-    serde_json::to_string(value).unwrap_or_default()
-}
-
-/// Char budget for a tool call's retained compact-JSON arguments.
-pub(crate) const INPUT_BUDGET: usize = 4_096;
-/// Char budget for a tool result's retained output (outputs are already
-/// capped upstream by the tools; this bounds transcript memory).
-pub(crate) const OUTPUT_BUDGET: usize = 16_384;
-
-/// Middle-out char cap preserving head and tail (first error + final summary
-/// both matter), on char boundaries.
-fn cap_middle(text: &str, budget: usize) -> String {
-    cap_middle_with(text, budget, "\n[… truncated …]\n")
-}
-
-/// [`cap_middle`] with a caller-chosen elision marker. Slices at
-/// `char_indices` boundaries instead of materializing a `Vec<char>`, so a
-/// multi-megabyte payload costs no allocation beyond the capped result.
-fn cap_middle_with(text: &str, budget: usize, marker: &str) -> String {
-    // Byte length bounds char count, so an in-budget payload returns without
-    // scanning; an over-budget one probes just past the boundary instead.
-    if text.len() <= budget || text.char_indices().nth(budget).is_none() {
-        return text.to_string();
-    }
-    let keep = budget.saturating_sub(marker.chars().count());
-    let head = keep / 2;
-    let tail = keep - head;
-    let head_end = text.char_indices().nth(head).map_or(text.len(), |(i, _)| i);
-    let tail_start = if tail == 0 {
-        text.len()
-    } else {
-        text.char_indices().nth_back(tail - 1).map_or(0, |(i, _)| i)
-    };
-    format!("{}{marker}{}", &text[..head_end], &text[tail_start..])
-}
-
-/// Per-leaf caps for [`cap_input_json`]: generous enough to keep any one
-/// argument readable, small enough that leaf capping alone usually lands the
-/// whole object under [`INPUT_BUDGET`].
-const INPUT_STR_CAP: usize = 512;
-const INPUT_ARR_CAP: usize = 32;
-
-/// Cap a tool call's retained arguments **inside** the JSON: long string
-/// leaves are middle-capped and oversized arrays elided, so the compact form
-/// stays *valid* JSON and ctrl+o can still pretty-print it. Only a
-/// pathological object that remains oversized after leaf capping falls back
-/// to the raw char cap (which the renderer shows as wrapped plain text).
-fn cap_input_json(value: &serde_json::Value, budget: usize) -> String {
-    let compact = compact_json(value);
-    if compact.len() <= budget {
-        return compact;
-    }
-    let mut capped = value.clone();
-    cap_json_leaves(&mut capped);
-    cap_middle(&compact_json(&capped), budget)
-}
-
-/// Recursively shrink the leaves of `value` in place (strings middle-capped
-/// on one line, arrays truncated with a `+N more` marker) without disturbing
-/// the object structure.
-fn cap_json_leaves(value: &mut serde_json::Value) {
-    match value {
-        serde_json::Value::String(s) => {
-            if s.len() > INPUT_STR_CAP {
-                *s = cap_middle_with(s, INPUT_STR_CAP, " […] ");
-            }
-        }
-        serde_json::Value::Array(items) => {
-            if items.len() > INPUT_ARR_CAP {
-                let dropped = items.len() - INPUT_ARR_CAP;
-                items.truncate(INPUT_ARR_CAP);
-                items.push(serde_json::Value::String(format!("[… +{dropped} more …]")));
-            }
-            for item in items.iter_mut() {
-                cap_json_leaves(item);
-            }
-        }
-        serde_json::Value::Object(map) => {
-            for item in map.values_mut() {
-                cap_json_leaves(item);
-            }
-        }
-        _ => {}
-    }
-}
-
-/// Format a tool-call input as a human-readable one-liner. Instead of raw
-/// JSON, this extracts the most relevant field(s) per tool name so the
-/// transcript reads naturally — `path` for file tools, `cmd` for shell, the
-/// query for search tools, and so on.
-fn format_tool_input(name: &str, input: &serde_json::Value) -> String {
-    let str_field = |key: &str| -> Option<String> {
-        input
-            .get(key)
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string())
-    };
-
-    // apply_edits carries its paths in a batch; surface them instead of the
-    // raw-JSON fallback so the transcript row reads like other file tools.
-    if name == "apply_edits"
-        && let Some(edits) = input.get("edits").and_then(|v| v.as_array())
-    {
-        let mut paths: Vec<&str> = Vec::new();
-        for edit in edits {
-            if let Some(p) = edit.get("path").and_then(|v| v.as_str())
-                && !paths.contains(&p)
-            {
-                paths.push(p);
-            }
-        }
-        if !paths.is_empty() {
-            let labels: Vec<String> = paths.iter().map(|p| truncate_field(p, 48)).collect();
-            return if paths.len() == 1 {
-                labels.into_iter().next().unwrap()
-            } else {
-                format!("{}  ({} files)", labels.join(", "), paths.len())
-            };
-        }
-    }
-
-    // Primary field per tool — the one the user cares about at a glance.
-    if let Some(p) = str_field("path").or_else(|| str_field("file_path")) {
-        return match name {
-            // Just the path. The old/new strings used to ride along here,
-            // truncated to 40 characters each — long enough to fill the row,
-            // never long enough to read, and describing exactly the change the
-            // inline diff below renders properly.
-            "edit_file" => p,
-            "write_file" => {
-                let lines = str_field("content").map(|c| c.lines().count()).unwrap_or(0);
-                format!("{p}  ({lines} lines)")
-            }
-            _ => p,
-        };
-    }
-
-    if let Some(cmd) = str_field("cmd").or_else(|| str_field("command")) {
-        return truncate_field(&cmd, 120);
-    }
-
-    if let Some(query) = str_field("query")
-        .or_else(|| str_field("pattern"))
-        .or_else(|| str_field("symbol"))
-    {
-        return truncate_field(&query, 80);
-    }
-
-    if let Some(prompt) = str_field("question").or_else(|| str_field("prompt")) {
-        return truncate_field(&prompt, 80);
-    }
-
-    // An argument-less call has nothing to summarize. `compact_json` renders
-    // `{}` for it, which the transcript then printed beside the tool name —
-    // and an empty object next to `project_overview` reads as an empty
-    // *result*, not as "this tool takes no arguments". Both this and the
-    // renderer treat the empty string as "no argument column"; saying it here
-    // as well is what keeps every consumer of `input` (the trace tab, the
-    // accessible renderer) agreeing about it.
-    if input.as_object().is_some_and(serde_json::Map::is_empty) || input.is_null() {
-        return String::new();
-    }
-    // Fallback: compact JSON, summarized.
-    summarize(&compact_json(input))
-}
-
-/// Truncate a field value to `max` chars with an ellipsis. Cuts *before*
-/// flattening: the newline→space replacement is one char in, one char out, so
-/// slicing the raw text first yields the identical result without copying a
-/// multi-megabyte `content`/`old_string` argument in full (the same reason
-/// [`cap_middle_with`] walks `char_indices` instead of materializing chars).
-fn truncate_field(s: &str, max: usize) -> String {
-    // `nth(max)` is `None` exactly when the text is `max` chars or shorter.
-    if s.char_indices().nth(max).is_none() {
-        return s.replace(['\n', '\r'], " ");
-    }
-    let head_end = s
-        .char_indices()
-        .nth(max.saturating_sub(1))
-        .map_or(s.len(), |(i, _)| i);
-    format!("{}…", s[..head_end].replace(['\n', '\r'], " "))
-}
-
-/// The workspace-relative path a file tool targets. Every built-in file tool
-/// (`read_file`/`write_file`/`edit_file`/`delete_file`) takes its path under
-/// the `path` key, and the engine emits `FileChange` for that same path — so
-/// this is the join key between a tool result and its diff.
-fn tool_input_path(input: &serde_json::Value) -> Option<String> {
-    if let Some(path) = input
-        .get("path")
-        .and_then(serde_json::Value::as_str)
-        .map(str::to_string)
-    {
-        return Some(path);
-    }
-    // apply_edits carries its paths in a batch, not at the top level; the
-    // first edit's path stands in so a single-file batch still renders an
-    // inline diff under its result row.
-    input
-        .get("edits")
-        .and_then(serde_json::Value::as_array)
-        .and_then(|edits| edits.first())
-        .and_then(|e| e.get("path"))
-        .and_then(serde_json::Value::as_str)
-        .map(str::to_string)
-}
-
-/// Whether a tool name is one of the file-*mutating* built-ins — the only
-/// tools whose result should carry an inline diff (reads must not). Must
-/// stay in lockstep with `file_change_of` in stella-cli's `command_deck.rs`,
-/// the `FileChange` emitter that owns this list.
-fn is_file_mutation(name: &str) -> bool {
-    matches!(
-        name,
-        "write_file" | "edit_file" | "apply_edits" | "delete_file"
-    )
-}
-
-/// Truncate a summary to [`SUMMARY_BUDGET`] chars with a middle-out elision —
-/// the head and tail both matter for a failing tool result (L-S3), so we keep
-/// both rather than head-truncating away the error tail.
-///
-/// Caps *before* flattening. `text` here is a raw tool payload that can be
-/// megabytes (the caller also passes it to [`cap_middle`]); replacing newlines
-/// first would copy the whole thing just to throw all but 200 chars away, and
-/// the replacement is one char in / one char out, so the two orders agree.
-fn summarize(text: &str) -> String {
-    cap_middle_with(text, SUMMARY_BUDGET, "...").replace(['\n', '\r'], " ")
 }
 
 #[cfg(test)]
